requirements.txt
# Core dependencies
google-adk>=0.1.0
google-genai>=0.3.0
python-dotenv>=1.0.0

# Additional utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
dateparser>=1.2.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0 